Better and safer handle for xidlehook

+++ .local/bin/inactivity.sh
@@ -1,9 +1,14 @@
 #!/bin/sh
 
-killall -q xidlehook
+startWatch() {
+    while pgrep -u $UID -x xidlehook > /dev/null; do sleep 1; done
+    if [ pgrep -c xidlehook -gt 1 ]
+        then
+            killall -q xidlehook
+    fi
+    exec xidlehook --not-when-fullscreen --not-when-audio --detect-sleep --timer 300 "lock.sh" - &
+}
 
-while pgrep -u $UID -x xidlehook > /dev/null; do sleep 1; done
-
-exec xidlehook --not-when-fullscreen --not-when-audio --detect-sleep --timer 300 "lock.sh" - &
+startWatch
 
 # $suspension --timer 600 "loginctl suspend" -
